pyproject.toml
[tool.poetry]
name = "forkwork"
version = "0.0.4"
description = "Find maintained forks of your favorite repositories"
authors = ["Andriy Orehov <andriyorehov@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/github-tooling/forkwork"

[tool.poetry.dependencies]
python = "^3.5"
click = "^7.0"
cachecontrol = "^0.12.5"
lockfile = "^0.12.2"
tabulate = "^0.8.3"
"github3.py" = "^1.3"
halo = "^0.0.26"
pendulum = "^2.0"

[tool.poetry.dev-dependencies]


[tool.poetry.scripts]
forkwork = 'forkwork.forkwork:cli'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
